feat: add wallet to new sidebar
