+++ apps/torus-portal/src/app/_components/portal-wallet-dropdown.tsx
@@ -1,0 +1,45 @@
+"use client";
+
+import type { SS58Address } from "@torus-network/sdk";
+
+import {
+  useCachedStakeOut,
+  useFreeBalance,
+} from "@torus-ts/query-provider/hooks";
+import { useTorus } from "@torus-ts/torus-provider";
+import { WalletDropdown } from "@torus-ts/ui/components/wallet-dropdown/wallet-dropdown";
+
+import { env } from "~/env";
+
+export function PortalWalletDropdown() {
+  const {
+    api,
+    accounts,
+    isInitialized,
+    selectedAccount,
+    handleLogout,
+    handleGetWallets,
+    handleSelectWallet,
+  } = useTorus();
+
+  const accountFreeBalance = useFreeBalance(
+    api,
+    selectedAccount?.address as SS58Address,
+  );
+
+  const stakeOut = useCachedStakeOut(env("NEXT_PUBLIC_TORUS_CACHE_URL"));
+
+  return (
+    <WalletDropdown
+      balance={accountFreeBalance.data}
+      stakeOut={stakeOut.data}
+      accounts={accounts}
+      isInitialized={isInitialized}
+      selectedAccount={selectedAccount}
+      handleLogout={handleLogout}
+      handleGetWallets={handleGetWallets}
+      handleSelectWallet={handleSelectWallet}
+      torusChainEnv={env("NEXT_PUBLIC_TORUS_CHAIN_ENV")}
+    />
+  );
+}
